// digitalbankingui/digital-banking-web/src/app/model/accounts.model.ts
export interface AccountDetails {
  accountId:            string;
  balance:              number;
  currentPage:          number;
  totalPages:           number;
  pageSize:             number;
  accountOperationDTOS: AccountOperation[];
}
export interface AccountOperation{
  id:number;
  operationDate: Date;
  amount:number;
  type:string;
  description:string;
}